Improvements to JPA annotations (e.g. length values set where appropriate) and raw data model start / stop coordinates changed to int values to be constistent with the post-processed model.
git-svn-id: https://interproscan.googlecode.com/svn/trunk@326 ea5da8b0-6d6b-11de-9de6-4d42d2ab87c0

+++ core/model/src/main/java/uk/ac/ebi/interpro/scan/model/raw/HamapRawMatch.java
@@ -15,7 +15,7 @@ public class HamapRawMatch extends ProfileScanRawMatch {
 
     public HamapRawMatch(String sequenceIdentifier, String model,
                          String signatureLibraryName, String signatureLibraryRelease,
-                         long locationStart, long locationEnd,
+                         int locationStart, int locationEnd,
                          double score, String generator) {
         super(sequenceIdentifier, model, signatureLibraryName, signatureLibraryRelease, locationStart, locationEnd, score, generator);
     }
